src/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic[email]==2.7.0
pydantic-settings==2.2.1
celery==5.3.6
redis==5.0.4
PyJWT==2.8.0
bcrypt==4.1.3
python-multipart==0.0.9
httpx==0.27.0
praw==7.7.1
google-generativeai==0.5.4
openai==1.30.1
anthropic==0.26.0
sentry-sdk[fastapi]==2.3.1
slowapi==0.1.9
python-dotenv==1.0.1
pytest==8.2.0
pytest-asyncio==0.23.6
aiohttp==3.9.5
structlog==24.1.0
tenacity==8.3.0
